pyproject.toml
[project]
name = "agent-coding-template"
version = "0.1.0"
description = "A coding agent used to bootstrap a repo with more complex instructions."
authors = [
  { name = "Khanh Nguyen", email = "khannguyen.qk@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"

dependencies = [
  "python-dotenv>=1.0.0",
  "langchain-community",
  "langchain-anthropic",
  "langgraph",
  "tavily-python",
  "langgraph-checkpoint-sqlite",
  "jinja2>=3.1.5"
]

[project.optional-dependencies]
dev = [
    "notebook",
    "ipykernel",
    "pytest>=7.0.0",
    "pytest-asyncio"
]
